requirements.txt
fastapi 
uvicorn 
google-cloud-language 
google-cloud-texttospeech 
google-cloud-vision 
google-generativeai 
python-dotenv
